+++ classes/observers/submission_observer.class.php
@@ -14,17 +14,19 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 /**
- * unicheck_event_group_submition.class.php
+ * submission_observer.class.php
  *
  * @package     plagiarism_unicheck
- * @author      Vadim Titov <v.titov@p1k.co.uk>
+ * @subpackage  plagiarism
+ * @author      Aleksandr Kostylev <a.kostylev@p1k.co.uk>
  * @copyright   UKU Group, LTD, https://www.unicheck.com
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-namespace plagiarism_unicheck\classes\event;
+namespace plagiarism_unicheck\classes\observers;
 
 use core\event\base;
+use plagiarism_unicheck\classes\entities\providers\unicheck_file_provider;
 use plagiarism_unicheck\classes\entities\unicheck_archive;
 use plagiarism_unicheck\classes\services\storage\unicheck_file_state;
 use plagiarism_unicheck\classes\unicheck_assign;
@@ -35,23 +37,68 @@ if (!defined('MOODLE_INTERNAL')) {
 }
 
 /**
- * Class unicheck_event_group_submition
+ * Class submission_observer
  *
  * @package     plagiarism_unicheck
  * @subpackage  plagiarism
- * @author      Vadim Titov <v.titov@p1k.co.uk>, Aleksandr Kostylev <a.kostylev@p1k.co.uk>
+ * @author      Aleksandr Kostylev <a.kostylev@p1k.co.uk>
  * @copyright   UKU Group, LTD, https://www.unicheck.com
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class unicheck_event_group_submition extends unicheck_abstract_event {
+class submission_observer extends abstract_observer {
+    /**
+     * DRAFT_STATUS
+     */
+    const DRAFT_STATUS = 'draft';
+
     /**
      * handle_event
      *
      * @param unicheck_core $core
      * @param base          $event
+     *
+     * @return bool
      */
-    public function handle_event(unicheck_core $core, base $event) {
+    public function status_updated(unicheck_core $core, base $event) {
 
+        global $DB;
+        if (!isset($event->other['newstatus'])) {
+            return false;
+        }
+        $newstatus = $event->other['newstatus'];
+        if (!$event->relateduserid) {
+            $core->enable_teamsubmission();
+        } else {
+            $core->userid = $event->relateduserid;
+        }
+
+        if ($newstatus == self::DRAFT_STATUS) {
+            $unfiles = \plagiarism_unicheck::get_area_files($event->contextid, UNICHECK_DEFAULT_FILES_AREA, $event->objectid);
+            $assignfiles = unicheck_assign::get_area_files($event->contextid, $event->objectid);
+
+            $files = array_merge($unfiles, $assignfiles);
+
+            $ids = [];
+            foreach ($files as $file) {
+                $plagiarismentity = $core->get_plagiarism_entity($file);
+                $internalfile = $plagiarismentity->get_internal_file();
+                $ids[] = $internalfile->id;
+            }
+
+            $allrecordssql = implode(',', $ids);
+            $DB->delete_records_select(UNICHECK_FILES_TABLE, "id IN ($allrecordssql) OR parent_id IN ($allrecordssql)");
+        }
+
+        return true;
+    }
+
+    /**
+     * status_viewed
+     *
+     * @param unicheck_core $core
+     * @param base          $event
+     */
+    public function status_viewed(unicheck_core $core, base $event) {
         $submission = unicheck_assign::get_user_submission_by_cmid($event->contextinstanceid);
         if (!$submission) {
             return;
@@ -60,7 +107,7 @@ class unicheck_event_group_submition extends unicheck_abstract_event {
         $assign = unicheck_assign::get($submission->assignment);
 
         /* Only for team submission */
-        if ($submission->status == unicheck_event_submission_updated::DRAFT_STATUS || !(bool)$assign->teamsubmission) {
+        if ($submission->status == self::DRAFT_STATUS || !(bool)$assign->teamsubmission) {
             return;
         }
 
@@ -80,15 +127,19 @@ class unicheck_event_group_submition extends unicheck_abstract_event {
                 continue;
             }
 
-            if (\plagiarism_unicheck::is_archive($assignfile)) {
-                $archive = new unicheck_archive($assignfile, $core);
-                $archive->upload();
+            try {
+                if (\plagiarism_unicheck::is_archive($assignfile)) {
+                    $archive = new unicheck_archive($assignfile, $core);
+                    $archive->upload();
 
-                continue;
-            }
+                    continue;
+                }
 
-            if ($internalfile->external_file_id == null) {
-                $this->add_after_handle_task($assignfile);
+                if ($internalfile->external_file_id == null) {
+                    $this->add_after_handle_task($assignfile);
+                }
+            } catch (\Exception $exception) {
+                unicheck_file_provider::to_error_state($internalfile, $exception->getMessage());
             }
         }