+++ tests/generator/lib.php
@@ -29,8 +29,6 @@ if (!defined('MOODLE_INTERNAL')) {
 
 global $CFG;
 
-require_once($CFG->dirroot . '/plagiarism/unicheck/classes/unicheck_core.class.php');
-
 use plagiarism_unicheck\classes\unicheck_core;
 
 /**