+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2018-04-18
+### Added
+- Self-plagiarism excluding
+- Add comments table
+- Add file metadata column
+- Permissions to change plugin settings
+- Show a notification if there is a limitation in the archive
+
+###Changed 
+- RAR,ZIP archive supporting is now OPTIONAL
+- Rename settings
+- Event observers
+- Settings descriptions
+- Sources for comparison display only those that are available on unicheck.com for the used API keys
+
+### Fixed
+- Cron crash when RAR or ZIP php extension is't installed
+
 ## [1.2.2] - 2018-04-11
 ### Fixed
 - Incorrect processing of large files
@@ -52,6 +70,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change validation mode
 - Massive refactoring comparing to Unplag plugin
 
+[1.3.0]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.3.0
 [1.2.2]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.2
 [1.2.1]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.1
 [1.2.0]: https://github.com/unicheck/moodle-plagiarism_unicheckcorp/releases/tag/v1.2.0